# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sistem Informasi Kepegawaian dan Status Kehadiran API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface Employee {
  id: number;
  nama: string;
  nopek: string;
  foto?: string | null;
  jabatan?: string | null;
  departemen?: string | null;
  createdAt: string;
}

export interface CreateEmployee {
  nama: string;
  nopek: string;
  foto?: string | null;
  jabatan?: string | null;
  departemen?: string | null;
}

export type AttendanceRecordStatus =
  (typeof AttendanceRecordStatus)[keyof typeof AttendanceRecordStatus];

export const AttendanceRecordStatus = {
  izin: "izin",
  tidak_hadir: "tidak_hadir",
  cuti: "cuti",
  dinas: "dinas",
} as const;

export interface AttendanceRecord {
  id: number;
  employeeId: number;
  employee?: Employee;
  status: AttendanceRecordStatus;
  tanggal: string;
  alasan?: string | null;
  keterangan?: string | null;
  createdAt: string;
}

export type CreateAttendanceRecordStatus =
  (typeof CreateAttendanceRecordStatus)[keyof typeof CreateAttendanceRecordStatus];

export const CreateAttendanceRecordStatus = {
  izin: "izin",
  tidak_hadir: "tidak_hadir",
  cuti: "cuti",
  dinas: "dinas",
} as const;

export interface CreateAttendanceRecord {
  employeeId: number;
  status: CreateAttendanceRecordStatus;
  tanggal: string;
  alasan?: string | null;
  keterangan?: string | null;
}

export interface AttendanceSummary {
  izin: number;
  tidak_hadir: number;
  cuti: number;
  dinas: number;
}

export type DocumentType = (typeof DocumentType)[keyof typeof DocumentType];

export const DocumentType = {
  SP3S: "SP3S",
  SIJ: "SIJ",
  CUTI: "CUTI",
  DINAS: "DINAS",
  SKMJ: "SKMJ",
  SURAT_TUGAS: "SURAT_TUGAS",
} as const;

export type DocumentStatus =
  (typeof DocumentStatus)[keyof typeof DocumentStatus];

export const DocumentStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface Document {
  id: number;
  employeeId: number;
  employee?: Employee;
  type: DocumentType;
  nomorSurat?: string | null;
  perihal?: string | null;
  tanggal: string;
  status: DocumentStatus;
  keterangan?: string | null;
  createdAt: string;
}

export type CreateDocumentType =
  (typeof CreateDocumentType)[keyof typeof CreateDocumentType];

export const CreateDocumentType = {
  SP3S: "SP3S",
  SIJ: "SIJ",
  CUTI: "CUTI",
  DINAS: "DINAS",
  SKMJ: "SKMJ",
  SURAT_TUGAS: "SURAT_TUGAS",
} as const;

export type CreateDocumentStatus =
  (typeof CreateDocumentStatus)[keyof typeof CreateDocumentStatus];

export const CreateDocumentStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface CreateDocument {
  employeeId: number;
  type: CreateDocumentType;
  nomorSurat?: string | null;
  perihal?: string | null;
  tanggal: string;
  status: CreateDocumentStatus;
  keterangan?: string | null;
}

export interface DocumentSummary {
  SP3S: number;
  SIJ: number;
  CUTI: number;
  DINAS: number;
  SKMJ: number;
  SURAT_TUGAS: number;
}

export type GetAttendanceRecordsParams = {
  status?: GetAttendanceRecordsStatus;
  date?: string;
};

export type GetAttendanceRecordsStatus =
  (typeof GetAttendanceRecordsStatus)[keyof typeof GetAttendanceRecordsStatus];

export const GetAttendanceRecordsStatus = {
  izin: "izin",
  tidak_hadir: "tidak_hadir",
  cuti: "cuti",
  dinas: "dinas",
} as const;

export type GetDocumentsParams = {
  type?: GetDocumentsType;
};

export type GetDocumentsType =
  (typeof GetDocumentsType)[keyof typeof GetDocumentsType];

export const GetDocumentsType = {
  SP3S: "SP3S",
  SIJ: "SIJ",
  CUTI: "CUTI",
  DINAS: "DINAS",
  SKMJ: "SKMJ",
  SURAT_TUGAS: "SURAT_TUGAS",
} as const;