# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sistem Informasi Kepegawaian dan Status Kehadiran API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /employees:
    get:
      operationId: getEmployees
      tags: [employees]
      summary: Get all employees
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create new employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployee"
      responses:
        "201":
          description: Employee created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get employee by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    put:
      operationId: updateEmployee
      tags: [employees]
      summary: Update employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployee"
      responses:
        "200":
          description: Employee updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    Employee:
      type: object
      properties:
        id:
          type: integer
        nama:
          type: string
        nopek:
          type: string
        foto:
          type: string
          nullable: true
        jabatan:
          type: string
          nullable: true
        departemen:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - nama
        - nopek
        - createdAt
    CreateEmployee:
      type: object
      properties:
        nama:
          type: string
        nopek:
          type: string
        foto:
          type: string
          nullable: true
        jabatan:
          type: string
          nullable: true
        departemen:
          type: string
          nullable: true
      required:
        - nama
        - nopek
...