# lib/api-zod/src/generated/types/employee.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sistem Informasi Kepegawaian dan Status Kehadiran API
 * OpenAPI spec version: 0.1.0
 */

export interface Employee {
  id: number;
  nama: string;
  nopek: string;
  foto?: string | null;
  jabatan?: string | null;
  departemen?: string | null;
  createdAt: string;
}